Replace RxJava2 with Coroutines
