+++ app/build.gradle
@@ -55,10 +55,6 @@ dependencies {
     def okHttpVersion = "3.9.1"
     implementation "com.squareup.okhttp3:logging-interceptor:$okHttpVersion"
 
-    //rx
-    implementation "io.reactivex.rxjava2:rxjava:2.2.10"
-    implementation "io.reactivex.rxjava2:rxandroid:2.0.1"
-
     //testing
     def junitVersion = "4.12"
     def mockitoVersion = "2.23.0"
@@ -71,4 +67,12 @@ dependencies {
     androidTestImplementation "com.android.support.test:rules:$androidTestVersion"
     implementation 'com.facebook.stetho:stetho:1.5.1'
     implementation 'com.facebook.stetho:stetho-okhttp3:1.5.1'
+
+    implementation "androidx.lifecycle:lifecycle-viewmodel-ktx:2.2.0"
+
+    // Coroutines
+    def coroutines_android_version = '1.3.2'
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_android_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_android_version"
+
 }
