--- app/src/main/java/com/n26/bitcointracker/utils/connectivity/ConnectivityUtil.kt
@@ -1,35 +1,0 @@
-package com.n26.bitcointracker.utils.connectivity
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-import android.os.Build
-
-
-object ConnectivityUtil {
-    @Suppress("DEPRECATION")
-    fun isNetworkAvailable(context: Context) : Boolean {
-        val cm = context.getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager?
-
-        if (cm != null) {
-            if (Build.VERSION.SDK_INT < 23) {
-                val ni = cm.activeNetworkInfo
-
-                if (ni != null) {
-                    return ni.isConnected && (ni.type == ConnectivityManager.TYPE_WIFI || ni.type == ConnectivityManager.TYPE_MOBILE)
-                }
-            } else {
-                val n = cm.activeNetwork
-
-                if (n != null) {
-                    val nc = cm.getNetworkCapabilities(n)
-
-                    return nc!!.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) || nc.hasTransport(
-                        NetworkCapabilities.TRANSPORT_WIFI
-                    )
-                }
-            }
-        }
-        return false
-    }
-}
